2048 Test Plan ========
Background
========

The objective of this game is to merge numbered tiles until you have one tile equal to 2048. The user manipulates the game by ‘tilting’ the board up, right, down, left in direction. If two tiles are the same number and are in line with each other are moved to a direction, which will allow a ‘merge’, a new tile is formed with the value of the total of the two tiles that merged. As the game progresses and as ‘tilts’ are performed, new tiles are created and positioned at random on the board. The user will merge as many tiles and as many times as needed to make a tile with the value 2048 WITHOUT filling the board with tiles and running out of possible merges as this will end the game short.

Specifications for running
========

For General use:For anyone wanting to play the game, with no other obligations.•	Web browser that allows JavaScript.You can also of course play this game via the command line with the appropriate files.For use of testing/changing program:For anyone wanting to look at how the program runs, test it, make or suggest improvements.•	GitHub Account•	IDE – Not Netbeans save yourself a headache•	C compiler and library – Mac OSX Mavericks…… What a joke.Testing
========

There are many aspects of this game that can be tested, such as:•	Movement of tiles•	Merging of tiles•	Game initialization•	Game end point (no more moves AND 2048 tile)•	Random placement of new valued tiles after each tilt•	Tiles merge with appropriate tiles (eg. same value)•	ScoringOther aspects, which effect the games functionality:•	Event handling (user input)•	Event recognition (user input)•	Action Listeners (user input)